fix: do checkstytle and fix
