+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+  "java.checkstyle.configuration": "${workspaceFolder}/config/checkstyle/checkstyle.xml"
+}
