--- src/main/java/cs2103w10/glados/enums/TaskType.java
@@ -1,7 +1,0 @@
-package cs2103w10.glados.enums;
-
-public enum TaskType {
-    DEADLINE,
-    EVENT,
-    TODO
-}
