--- src/main/java/cs2103w10/glados/exceptions/ArgumentsNotFoundException.java
@@ -1,10 +1,0 @@
-package cs2103w10.glados.exceptions;
-
-public class ArgumentsNotFoundException extends GladosException {
-    /**
-     * Constructs ArgumentsNotFoundException by calling the parent class.
-     */
-    public ArgumentsNotFoundException() {
-        super( "Unknown arguments. Please try again.");
-    }
-}
