--- src/main/java/cs2103w10/glados/exceptions/CommandNotFoundException.java
@@ -1,10 +1,0 @@
-package cs2103w10.glados.exceptions;
-
-public class CommandNotFoundException extends GladosException {
-    /**
-     * Constructs CommandNotFoundException by calling the parent class.
-     */
-    public CommandNotFoundException() {
-        super( "Unknown command. Please try again.");
-    }
-}
