--- src/main/java/cs2103w10/glados/exceptions/DescriptionNotFoundException.java
@@ -1,10 +1,0 @@
-package cs2103w10.glados.exceptions;
-
-public class DescriptionNotFoundException extends GladosException {
-    /**
-     * Constructs DescriptionNotFoundException by calling the parent class.
-     */
-    public DescriptionNotFoundException() {
-        super("Description for a task cannot be empty.");
-    }
-}
