--- src/main/java/cs2103w10/glados/exceptions/GladosException.java
@@ -1,10 +1,0 @@
-package cs2103w10.glados.exceptions;
-
-public abstract class GladosException extends Exception {
-    /**
-     * Constructs GladosException by calling the parent class.
-     */
-    public GladosException(String message) {
-        super(message);
-    }
-}
