--- src/main/java/cs2103w10/glados/tasks/Todo.java
@@ -1,59 +1,0 @@
-package cs2103w10.glados.tasks;
-
-import cs2103w10.glados.enums.TaskType;
-
-public class Todo extends Task{
-
-    /**
-     * Constructs new Todo object by calling parent.
-     * 
-     * @param description Description of Todo.
-     */
-    public Todo(String description) {
-        super(description);
-    }
-
-    /**
-     * {@inheritDoc}
-     * Returns Todo task type icon.
-     * 
-     * @return [T] icon.
-     */
-    @Override
-    public String getTaskTypeIcon() {
-        return "[T]";
-    }
-
-    /**
-     * {@inheritDoc}
-     * Returns Todo task type.
-     * 
-     * @return Todo enum.
-     */
-    @Override
-    public TaskType getTaskType() {
-        return TaskType.TODO;
-    }
-
-    /**
-     * {@inheritDoc}
-     * Returns Todo description.
-     * 
-     * @return Todo string.
-     */
-    @Override
-    public String getDescription() {
-        return this.description;
-    }
-
-    /**
-     * {@inheritDoc}
-     * Returns Todo string for saving into data folder.
-     * 
-     * @return Todo save format string.
-     */
-    @Override
-    public String getSaveFormat() {
-        return this.description;
-    }
-}
