+++ src/main/java/cstwooneohthree/glados/enums/TaskType.java
@@ -1,0 +1,12 @@
+package cstwooneohthree.glados.enums;
+
+/**
+ * Enums for different ask types.
+ *
+ * @author jayjay19630
+ */
+public enum TaskType {
+    DEADLINE,
+    EVENT,
+    TODO
+}
