+++ src/main/java/cstwooneohthree/glados/exceptions/CommandNotFoundException.java
@@ -1,0 +1,15 @@
+package cstwooneohthree.glados.exceptions;
+
+/**
+ * CommandNotFoundException class when command does not exist.
+ *
+ * @author jayjay19630
+ */
+public class CommandNotFoundException extends GladosException {
+    /**
+     * Constructs CommandNotFoundException by calling the parent class.
+     */
+    public CommandNotFoundException() {
+        super("Unknown command. Please try again.");
+    }
+}
